Fix new users bug
The bug was caused by the absence of stations with newly created
accounts. The workaround is to automatically create a station, this way
we don't have to handle the unusual case with no station.
There is still to prevent the deletion of the last station, this will
come in another commit.