add JVM warning message
It seem that Oracle JVM has a bug in the sound stack that occurs in case
of buffer underflows of the dataline. A workaround would be to keep
sending blank frames to prevent the underflow to occur (slow network or
pause).
So a message will be displayed when Beluga Player is started with Oracle
JVM under GNU/Linux.